% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGDB-package.R
\docType{package}
\name{FGDB-package}
\alias{FGDB}
\alias{FGDB-package}
\title{FGDB: FluvialGeomorph Database}
\description{
Tools and relational schemas for validating, loading, and managing FluvialGeomorph data in the FluvialGeomorph Database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FluvialGeomorph/FGDB}
  \item Report bugs at \url{https://github.com/FluvialGeomorph/FGDB/issues}
}

}
\author{
\strong{Maintainer}: Michael Dougherty \email{Michael.P.Dougherty@usace.army.mil} (\href{https://orcid.org/0000-0002-1465-5927}{ORCID})

Authors:
\itemize{
  \item Michael Dougherty \email{Michael.P.Dougherty@usace.army.mil} (\href{https://orcid.org/0000-0002-1465-5927}{ORCID})
}

}
\keyword{internal}
